chore(offlinedocs): rewrite `next` based docs to `vite`
