+++ offlinedocs/.gitignore
@@ -1,5 +1,3 @@
-# See https://help.github.com/articles/ignoring-files/ for more about ignoring files.
-
 # dependencies
 /node_modules
 /.pnp
@@ -8,12 +6,12 @@
 # testing
 /coverage
 
-# next.js
-/.next/
-/out/
-
 # production
+/out/
 /build
+
+# react router
+/.react-router/
 
 # misc
 .DS_Store
@@ -25,9 +23,6 @@ pnpm-debug.log*
 
 # local env files
 .env*.local
-
-# vercel
-.vercel
 
 # typescript
 *.tsbuildinfo
