+++ offlinedocs/index.html
@@ -1,0 +1,11 @@
+<!doctype html>
+<html lang="en">
+	<head>
+		<meta charset="UTF-8" />
+		<meta name="viewport" content="width=device-width, initial-scale=1.0" />
+		<title>Coder Docs</title>
+	</head>
+	<body>
+		<script type="module" src="/src/entry.client.tsx"></script>
+	</body>
+</html>
