--- offlinedocs/next.config.js
@@ -1,8 +1,0 @@
-/** @type {import('next').NextConfig} */
-const nextConfig = {
-	output: "export",
-	reactStrictMode: true,
-	trailingSlash: true,
-};
-
-module.exports = nextConfig;
